Added attribute ref to JWT
The ref attribute contains the http referer header and can be used to bind the jwt to a specific host.. The API gateway can then compare the value with the actual http referer to make sure they are the same. Also, the hostname part of the referer can be used to qualify users during access control checks to make sure the user has the required permissions at the specific host.